organização dos componentes e alteração nas cores do labetube
