+++ semana5/labetube/src/components/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Footer() {
+  return (
+    <footer>
+      <h4>Oi! Eu moro no footer!</h4>
+    </footer>
+  );
+}
+
+export default Footer;
