+++ semana5/labetube/src/components/Nav.js
@@ -1,0 +1,19 @@
+import React from 'react';
+//import './style.css';
+
+function Nav() {
+  return (
+    <nav className='menu-vertical'>
+      <ul>
+        <li className='botoes-meunu-vertical'>Início</li>
+        <li className='botoes-meunu-vertical'>Em alta</li>
+        <li className='botoes-meunu-vertical'>Inscrições</li>
+        <hr />
+        <li className='botoes-meunu-vertical'>Originais</li>
+        <li className='botoes-meunu-vertical'>Histórico</li>
+      </ul>
+    </nav>
+  );
+}
+
+export default Nav;
