+++ semana5/labetube/src/components/Search.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Search() {
+  return (
+    <header>
+      <h1>Lab Tube</h1>
+      <input type='text' placeholder='Busca' id='campoDeBusca' />
+    </header>
+  );
+}
+
+export default Search;
